Isolate test that does reload to fix cross-talk failure
